feat: реализация test_auth_admin и перенос тестов поиска в новую структуру
