+++ backend/tests/conftest.py
@@ -1,0 +1,75 @@
+"""Конфигурация pytest для тестов."""
+
+import asyncio
+from pathlib import Path
+from typing import AsyncGenerator, Generator
+
+import pytest
+import pytest_asyncio
+from httpx import ASGITransport, AsyncClient
+from sqlalchemy.ext.asyncio import AsyncSession, create_async_engine
+from sqlalchemy.orm import sessionmaker
+from sqlalchemy.pool import NullPool
+
+from backend.core.db import Base, get_session
+from backend.main import app
+from backend.tests.fixtures import admin_user, menu_items_fixture, user_free, user_premium
+
+
+BASE_DIR = Path(__file__).resolve(strict=True).parent.parent
+
+TEST_DB = BASE_DIR / "test.db"
+SQLALCHEMY_DATABASE_URL = f"sqlite+aiosqlite:///{str(TEST_DB)}"
+
+test_engine = create_async_engine(
+    SQLALCHEMY_DATABASE_URL,
+    echo=False,
+    poolclass=NullPool,
+    connect_args={"check_same_thread": False},
+)
+
+TestSessionLocal = sessionmaker(
+    test_engine,
+    class_=AsyncSession,
+    expire_on_commit=False,
+)
+
+
+@pytest.fixture(scope="session")
+def event_loop() -> Generator:
+    """Создать event loop для всей сессии тестов."""
+    loop = asyncio.get_event_loop_policy().new_event_loop()
+    yield loop
+    loop.close()
+
+
+@pytest_asyncio.fixture(scope="function")
+async def db() -> AsyncGenerator[AsyncSession, None]:
+    """Создать сессию БД для каждого теста."""
+    async with test_engine.begin() as conn:
+        await conn.run_sync(Base.metadata.create_all)
+    async with TestSessionLocal() as session:
+        yield session
+    async with test_engine.begin() as conn:
+        await conn.run_sync(Base.metadata.drop_all)
+
+
+@pytest_asyncio.fixture(scope="function")
+async def async_client(db: AsyncSession) -> AsyncGenerator[AsyncClient, None]:
+    """Создать async HTTP клиент для тестирования API."""
+
+    async def override_get_session():
+        yield db
+
+    app.dependency_overrides[get_session] = override_get_session
+
+    transport = ASGITransport(app=app)
+    async with AsyncClient(transport=transport, base_url="http://test") as client:
+        yield client
+    app.dependency_overrides.clear()
+
+
+@pytest.fixture
+def anyio_backend():
+    """Указать backend для anyio (используется pytest-asyncio)."""
+    return "asyncio"
