+++ backend/tests/test_search_functionality.py
