+++ pyproject.toml
@@ -57,6 +57,7 @@ httpx = "^0.28.1"
 pytest-mock = "^3.15.1"
 aiosqlite = "^0.21.0"
 greenlet = "^3.2.4"
+pytest-cov = "^7.0.0"
 
 [build-system]
 requires = ["poetry-core"]
@@ -83,3 +84,6 @@ skip-magic-trailing-comma = false
 
 [tool.ruff.lint.pydocstyle]
 convention = "google"
+
+[tool.pytest.ini_options]
+asyncio_mode = "auto"
